who knows, I think i fixed the seed data now

+++ app/controllers/friendships_controller.rb
@@ -63,8 +63,10 @@ class FriendshipsController < ApplicationController
 
 	def destroy
 		friendship = @current_user.friendships.find(params[:id])
-		inverse_friendship = Friendship.find_by(friend_id: @current_user.id, user_id: friendship.friend_id)
-		inverse_friendship.destroy
+		if Friendship.where(friend_id: @current_user.id, user_id: friendship.friend_id)
+			inverse_friendship = Friendship.find_by(friend_id: @current_user.id, user_id: friendship.friend_id)
+			inverse_friendship.destroy
+		end
 		friendship.destroy
 		flash[:notice] = "Successfully deleted friendship."
 		redirect_to friendships_path
